change loop variables type

+++ contract/contracts/Bingo.sol
@@ -80,8 +80,8 @@ contract Bingo {
 
     function getNumbers() public pure returns (uint8[] memory) {
         uint8[] memory numbers = new uint8[](75);
-        for (uint i = 0; i < 75; i++) {
-            numbers[i] = uint8(i + 1);
+        for (uint8 i = 0; i < 75; i++) {
+            numbers[i] = i + 1;
         }
         return numbers;
     }
